Create a dedicated content list transformer
