+++ content/webapp/services/prismic/transformers/index.ts
@@ -40,6 +40,7 @@ import {
   getWeight,
   transformCollectionVenueSlice,
   transformContactSlice,
+  transformContentListSlice,
   transformDeprecatedImageListSlice,
   transformDiscussionSlice,
   transformEditorialImageGallerySlice,
@@ -178,10 +179,6 @@ type PromoImage = {
   superWidescreenImage?: ImageType;
 };
 
-// TODO: Consider moving this into a dedicated file for body transformers.
-// TODO: Rather than doing transformation inline, have this function consistently
-// call out to other transformer functions (a la contentList).
-// See https://github.com/wellcomecollection/wellcomecollection.org/pull/7679/files#r811138079
 export function transformBody(body: Body): BodyType {
   return body
     .map(slice => {
@@ -205,48 +202,7 @@ export function transformBody(body: Body): BodyType {
           return transformTitledTextListSlice(slice);
 
         case 'contentList':
-          type ContentListPrismicDocument =
-            | MultiContentPrismicDocument
-            | GuidePrismicDocument
-            | SeasonPrismicDocument
-            | CardPrismicDocument;
-
-          const contents: ContentListPrismicDocument[] = slice.items
-            .map(item => item.content)
-            .filter(isFilledLinkToDocumentWithData);
-
-          return {
-            type: 'contentList',
-            weight: getWeight(slice.slice_label),
-            value: {
-              title: asText(slice.primary.title),
-              // TODO: The old code would look up a `hasFeatured` field on `slice.primary`,
-              // but that doesn't exist in our Prismic model.
-              // hasFeatured: slice.primary.hasFeatured,
-              items: contents
-                .map(content => {
-                  switch (content.type) {
-                    case 'pages':
-                      return transformPage(content);
-                    case 'guides':
-                      return transformGuide(content);
-                    case 'event-series':
-                      return transformEventSeries(content);
-                    case 'exhibitions':
-                      return transformExhibition(content);
-                    case 'articles':
-                      return transformArticle(content);
-                    case 'events':
-                      return transformEvent(content);
-                    case 'seasons':
-                      return transformSeason(content);
-                    case 'card':
-                      return transformCard(content);
-                  }
-                })
-                .filter(Boolean),
-            },
-          };
+          return transformContentListSlice(slice);
 
         case 'collectionVenue':
           return transformCollectionVenueSlice(slice);
