+++ content/webapp/types/multi-content.ts
@@ -8,14 +8,20 @@ import { Series } from './series';
 import { Guide } from './guides';
 import { Weblink } from './weblinks';
 import { Project } from './projects';
+import { ArticleScheduleItem } from './article-schedule-items';
+import { Card } from './card';
+import { Season } from './seasons';
 
 export type MultiContent =
+  | ArticleScheduleItem
+  | Card
   | Page
   | EventSeries
   | Book
   | Event
   | Article
   | Exhibition
+  | Season
   | Series
   | Guide
   | Weblink
